refactor dtos directories
