+++ Dto/Responses/PseudoNoContentResponseDto.cs
@@ -1,7 +1,7 @@
 ﻿using System.Text.Json.Serialization;
 
 
-namespace asp_net_po_schedule_management_server.Dto.Misc
+namespace asp_net_po_schedule_management_server.Dto.Responses
 {
     public sealed class PseudoNoContentResponseDto
     {
