+++ Dto/Shared/DictionaryHashInjectionFragment.cs
@@ -1,6 +1,6 @@
 ﻿namespace asp_net_po_schedule_management_server.Dto.Shared
 {
-    public abstract class ResponseDictionaryHashInjection
+    public abstract class DictionaryHashInjectionFragment
     {
         public string DictionaryHash { get; set; }
     }
